Initial deliverables uploaded for beta.
Beta files for check-in. The futti-rules.pdf document goes over the rules of the game. There are two boards needed to play, this are the futti-doubles-chart.pdf and the futti-main-gameboard.pdf. There are two sets of player cards that you can use to jumpstart the game and create your own teams. The newer cards have a better look and feel, but it might be your preference to use the older ones. The zip file contains everything you need to play.